Project restructure. Added all tests to /__tests__
