--- src/components/TwitterFeed/TwitterFeed.test.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { render, cleanup } from "@testing-library/react";
-import "@testing-library/jest-dom/extend-expect";
-import VirtualizedList from '../TwitterFeed/TwitterFeed';
-
-describe("TwitterFeed Component", () => {
-    it("renders the listItem", () => {
-        
-    });
-});
